+++ src/keybindings/cutWord.ts
@@ -4,41 +4,34 @@ import { debug, getNumber, getSettings, resetNumber } from "@/common/funcs";
 export default (logseq: ILSPluginUser) => {
   const settings = getSettings();
 
-  const bindings = Array.isArray(settings.increase)
-    ? settings.increase
-    : [settings.increase];
+  const bindings = Array.isArray(settings.cutWord)
+    ? settings.cutWord
+    : [settings.cutWord];
 
   bindings.forEach((binding, index) => {
     logseq.App.registerCommandPalette(
       {
-        key: "vim-shortcut-increase-" + index,
-        label: "Increase",
+        key: "vim-shortcut-cut-word-" + index,
+        label: "Cut word",
         keybinding: {
           mode: "non-editing",
           binding,
         },
       },
       async () => {
-        debug("increase");
-
-        const number = getNumber();
-        resetNumber();
+        debug("cut word");
 
         const selected = await logseq.Editor.getSelectedBlocks();
         if (selected.length > 1) {
           for (let block of selected) {
-            const content = block.content.replace(/-?[0-9]+/, (match, p1) => {
-              return `${parseInt(match) + number}`;
-            });
+            const content = block.content.split(" ").slice(1).join(" ");
 
             await logseq.Editor.updateBlock(block.uuid, content);
           }
         } else {
           const block = await logseq.Editor.getCurrentBlock();
           if (block) {
-            const content = block.content.replace(/-?[0-9]+/, (match, p1) => {
-              return `${parseInt(match) + number}`;
-            });
+            const content = block.content.split(" ").slice(1).join(" ");
 
             await logseq.Editor.updateBlock(block.uuid, content);
           }